+++ src/Billing/Controllers/PayPalController.cs
@@ -1,6 +1,7 @@
 ﻿using System.Text;
 using Bit.Billing.Models;
 using Bit.Core.AdminConsole.Repositories;
+using Bit.Core.Billing.Premium.Services;
 using Bit.Core.Billing.Services;
 using Bit.Core.Entities;
 using Bit.Core.Enums;
