+++ src/Core/Billing/Extensions/ServiceCollectionExtensions.cs
@@ -1,6 +1,7 @@
 ﻿using Bit.Core.Billing.Licenses.Extensions;
 using Bit.Core.Billing.Organizations.Services;
 using Bit.Core.Billing.Payment.Services;
+using Bit.Core.Billing.Premium.Services;
 using Bit.Core.Billing.Pricing;
 using Bit.Core.Billing.Services;
 using Bit.Core.Billing.Services.Implementations;
