+++ src/Core/Billing/Premium/Requests/PreviewIndividualInvoiceRequestModel.cs
@@ -1,6 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
+using Bit.Core.Billing.Models.Api.Requests;
 
-namespace Bit.Core.Billing.Models.Api.Requests.Accounts;
+namespace Bit.Core.Billing.Premium.Requests;
 
 public class PreviewIndividualInvoiceRequestBody
 {
