+++ src/Core/Services/Implementations/StripePaymentService.cs
@@ -8,6 +8,7 @@ using Bit.Core.Billing.Models.Api.Responses;
 using Bit.Core.Billing.Organizations.Models;
 using Bit.Core.Billing.Organizations.Requests;
 using Bit.Core.Billing.Payment.Models;
+using Bit.Core.Billing.Premium.Requests;
 using Bit.Core.Billing.Pricing;
 using Bit.Core.Billing.Services;
 using Bit.Core.Billing.Services.Contracts;
