+++ src/Core/Services/Implementations/UserService.cs
@@ -14,6 +14,8 @@ using Bit.Core.Auth.Models.Business.Tokenables;
 using Bit.Core.Billing.Models;
 using Bit.Core.Billing.Models.Sales;
 using Bit.Core.Billing.Payment.Models;
+using Bit.Core.Billing.Premium.Models;
+using Bit.Core.Billing.Premium.Services;
 using Bit.Core.Billing.Services;
 using Bit.Core.Context;
 using Bit.Core.Entities;
